Allow skipping step-cli installation

+++ scripts/setup-devtools.sh
@@ -70,16 +70,18 @@ if [ "${ENABLE_DEVTOOLS}" != "false" ]; then
         *) echo "unsupported architecture"; exit 1 ;;
     esac
 
-    curl -sSL -o /tmp/step-cli.tar.gz https://github.com/smallstep/cli/releases/download/v${STEP_CLI_VERSION}/step_linux_${STEP_CLI_VERSION}_${ARCHITECTURE}.tar.gz
-    curl -sSL -o /tmp/SHASUMS256.txt https://github.com/smallstep/cli/releases/download/v${STEP_CLI_VERSION}/checksums.txt
+    if [ "${STEP_CLI_VERSION}" != "none" ]; then
+        curl -sSL -o /tmp/step-cli.tar.gz https://github.com/smallstep/cli/releases/download/v${STEP_CLI_VERSION}/step_linux_${STEP_CLI_VERSION}_${ARCHITECTURE}.tar.gz
+        curl -sSL -o /tmp/SHASUMS256.txt https://github.com/smallstep/cli/releases/download/v${STEP_CLI_VERSION}/checksums.txt
 
-    cat /tmp/SHASUMS256.txt | grep "$(sha256sum /tmp/step-cli.tar.gz | cut -d ' ' -f 1)"
+        cat /tmp/SHASUMS256.txt | grep "$(sha256sum /tmp/step-cli.tar.gz | cut -d ' ' -f 1)"
 
-    mkdir -p /tmp/step-cli
-    tar -xz -f /tmp/step-cli.tar.gz -C /tmp/step-cli --strip-components=1
-    cp -v /tmp/step-cli/bin/step /usr/local/bin/step
+        mkdir -p /tmp/step-cli
+        tar -xz -f /tmp/step-cli.tar.gz -C /tmp/step-cli --strip-components=1
+        cp -v /tmp/step-cli/bin/step /usr/local/bin/step
 
-    rm -rf /tmp/step-cli /tmp/SHASUMS256.txt /tmp/step-cli.tar.gz
+        rm -rf /tmp/step-cli /tmp/SHASUMS256.txt /tmp/step-cli.tar.gz
+    esac
 
     echo "Done!"
 fi
